Split SafeCmd run tests by concern
test_safe_cmd_run.py carried baseline execution, output-option, and
timeout/cancellation tests in one module. Split them into
test_safe_cmd_output.py (capture, echo, sinks, decoding, encode-failure
containment) and test_safe_cmd_timeout.py (TimeoutExpired contract,
cancellation escalation, public-boundary non-positive timeouts), leaving
only baseline execution coverage in test_safe_cmd_run.py.

Each module keeps its local run()/run_sync() parametrization
(_execute_async/_execute_sync, execution_strategy, python_builder),
matching the existing pattern in test_safe_cmd_stdin.py and
test_safe_cmd_context.py. No assertions, names, or semantics changed;
tests/helpers/execution.py remains a type/helper module without
fixtures. The maturin wheel snapshot is regenerated for the two new
shipped test files.
